# docs/aidlc/05-api-specification/Fantasy EPL League Manager — OpenAPI Specification v1.0.yaml
openapi: 3.0.3
info:
  title: Fantasy EPL League Manager API
  version: "1.0"
  description: |
    # Fantasy EPL League Manager
    ## OpenAPI Specification — Version 1.0

    **Document Status:** Baseline API Specification

    **Inputs:**
    - Architecture and Domain Model v1.15 (`Fantasy EPL League Manager — Architecture and Domain Model v1.15.md`) — Section 9's endpoint sketches (§9.2), the aggregate/entity model (§6), logical data architecture (§8), and cross-cutting API conventions (§9.3) that this document turns into concrete request/response DTOs and validation rules.
    - Business Requirements Document v1.17 (`Fantasy EPL League Manager — Business Requirements Document v1.17.md`) — authoritative business rules, cited throughout by `BR-###`.
    - Epic and Feature Backlog v1.10 (`Fantasy EPL League Manager — Epic and Feature Backlog v1.10.md`) — feature IDs (`F-###.#`) cited on each operation group to show which feature it serves and in what build-sequence position (Backlog §5).

    **Purpose:** Turn Architecture §9's illustrative endpoint list into a complete, implementation-ready OpenAPI contract — every resource, request/response schema, validation rule, and error shape needed to start writing controllers and client code. This is the artifact Architecture §16 lists as recommended-next-artifact #3.

...

...

    ## Conventions (Architecture §9.1/§9.3)

    - REST over HTTPS only; JSON request/response bodies.
...

    ## Version History

    ### Version 1.0

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /leagues:
    post:
      tags: [LeagueSeason]
      operationId: createLeague
      summary: Create a League (F-003.1, BR-023, BR-024, BR-026)
      description: The creator's `LeagueMembership` is created with `IsAdministrator = true` (BR-024).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLeagueRequest'
      responses:
        '201':
          description: League created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
        '400':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [LeagueSeason]
      operationId: listMyLeagues
      summary: List Leagues the caller has an active LeagueMembership in (BR-030, BR-031)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/League'
  /leagues/{leagueId}:
    parameters:
      - $ref: '#/components/parameters/leagueId'
    get:
      tags: [LeagueSeason]
      operationId: getLeague
      summary: Get League details
      x-authorization: Caller must be an active member of leagueId.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [LeagueSeason]
      operationId: updateLeague
      summary: Update League name/description/status (League Administrator only, BR-025)
      x-authorization: Caller must be the League Administrator of leagueId.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeagueRequest'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /leagues/{leagueId}/invitations:
    parameters:
      - $ref: '#/components/parameters/leagueId'
    post:
      tags: [LeagueSeason]
      operationId: createInvitation
      summary: Issue a League invitation (F-003.2, BR-027–BR-029)
      x-authorization: Caller must be the League Administrator of leagueId.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: Invitation created; expires in `LeagueConfiguration.InvitationExpiration` (default 7 days, BR-029).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags: [LeagueSeason]
      operationId: listInvitations
      summary: List pending/expired invitations for a League (League Administrator only)
      x-authorization: Caller must be the League Administrator of leagueId.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invitation'
  /leagues/{leagueId}/invitations/{invitationId}:
    delete:
      tags: [LeagueSeason]
      operationId: revokeInvitation
      summary: Revoke a pending invitation (League Administrator only)
      x-authorization: Caller must be the League Administrator of leagueId.
      parameters:
        - $ref: '#/components/parameters/leagueId'
        - name: invitationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Revoked.
  /invitations/{token}/accept:
    post:
      tags: [LeagueSeason]
      operationId: acceptInvitation
      summary: Accept a League invitation (F-003.2, BR-027)
      description: Creates or reactivates the caller's LeagueMembership. Fails with 410 if the token is expired (BR-029) or already used/revoked.
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Membership created/reactivated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueMembership'
        '410':
          description: Invitation expired, already accepted, or revoked (BR-029).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /epl/players:
    get:
      tags: [PlayerData]
      operationId: listPlayers
      summary: List EPL Players, optionally filtered by club (F-004.1, BR-227, BR-228)
      parameters:
        - name: clubId
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/positionFilter'
        - $ref: '#/components/parameters/searchQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
  /epl/players/{playerId}:
    get:
      tags: [PlayerData]
      operationId: getPlayer
      summary: Get one Player's reference data
      parameters:
        - name: playerId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          $ref: '#/components/responses/NotFound'
  /epl/gameweeks:
    get:
      tags: [PlayerData]
      operationId: listGameweeks
      summary: List Gameweeks for an EPL Season (F-004.2, BR-092, BR-229, BR-230)
      parameters:
        - name: eplSeasonIdentifier
          in: query
          required: true
          schema: { type: string, example: "2026/27" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gameweek'
  /epl/gameweeks/{gameweekId}/fixtures:
    get:
      tags: [PlayerData]
      operationId: getGameweekFixtures
      summary: Fixtures for a Gameweek (F-004.6/F-007.1, BR-332, BR-333, BR-336)
      description: Also the read reused verbatim by the Weekly Roster screen's Gameweek schedule display (BR-336).
      parameters:
        - name: gameweekId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fixture'
  /epl/seasons/{eplSeasonId}/table:
    get:
      tags: [PlayerData]
      operationId: getEplTable
      summary: Real-world EPL league table for a Season (F-004.6, BR-329–BR-331)
      description: Distinct from the Fantasy League's own Standings (`/leagues/{leagueId}/seasons/{seasonId}/standings`). Read-only; no Administrator-override path (BR-334).
      parameters:
        - name: eplSeasonId
          in: path
          required: true
          schema: { type: string, example: "2026/27" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClubStanding'

  # ──────────────────────────────────────────── Draft ────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    leagueId:
      name: leagueId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    positionFilter:
      name: position
      in: query
      description: Filter by position (BR-310, BR-314).
      schema:
        $ref: '#/components/schemas/Position'
    searchQuery:
      name: search
      in: query
      description: Case-insensitive substring match on player name (BR-311, BR-315).
      schema: { type: string }
...
  responses:
    ValidationError:
      description: Request failed shape/format validation (BR-166).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Forbidden:
      description: Authenticated, but the caller fails the operation's object-level authorization check (BR-163, BR-254).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource does not exist, or exists but the caller has no visibility into it (never distinguished from a plain 404, to avoid leaking existence).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      description: RFC 7807-style error body (Architecture §9.3), extended with a stable machine-readable errorCode.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errorCode:
          type: string
          example: player_already_owned
        correlationId: { type: string, description: Propagated into logs (BR-241). }
    Username:
      type: string
      minLength: 3
      maxLength: 24
      pattern: '^[A-Za-z0-9_]+$'
      description: Unique among active Users (BR-004); reused screens display it verbatim (BR-003). Length/charset are an implementation default, not yet a cited BR — confirm before locking a physical CHECK constraint.
...
    LeagueStatus:
      type: string
      enum: [Active, Archived]
    League:
      type: object
      properties:
        leagueId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 1000 }
        status:
          $ref: '#/components/schemas/LeagueStatus'
        createdByMembershipId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    CreateLeagueRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 1000 }
    UpdateLeagueRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 1000 }
        status:
          $ref: '#/components/schemas/LeagueStatus'
    MembershipStatus:
      type: string
      enum: [Invited, Active, Left]
    LeagueMembership:
      type: object
      properties:
        leagueMembershipId: { type: string, format: uuid }
        leagueId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        username:
          $ref: '#/components/schemas/Username'
        isAdministrator: { type: boolean }
        leagueIconId: { type: string, format: uuid, nullable: true }
        effectiveIconId: { type: string, format: uuid, description: "leagueIconId if set, else the User's default icon (BR-008)." }
        status:
          $ref: '#/components/schemas/MembershipStatus'
        joinedAt: { type: string, format: date-time }
        leftAt: { type: string, format: date-time, nullable: true }
    InvitationStatus:
      type: string
      enum: [Pending, Accepted, Expired, Revoked]
    Invitation:
      type: object
      properties:
        invitationId: { type: string, format: uuid }
        leagueId: { type: string, format: uuid }
        seasonId: { type: string, format: uuid, nullable: true }
        destination: { type: string, description: Email or phone number (BR-028). }
        status:
          $ref: '#/components/schemas/InvitationStatus'
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    CreateInvitationRequest:
      type: object
      required: [destination, channel]
      properties:
        destination: { type: string }
        channel:
          type: string
          enum: [Email, Sms]
        seasonId: { type: string, format: uuid, nullable: true }
...
    Position:
      type: string
      enum: [Gk, Def, Mid, Fwd]
...
    Player:
      type: object
      properties:
        playerId: { type: string, format: uuid }
        eplPlayerId: { type: string, description: External identifier used for idempotent sync upserts (BR-232). }
        name: { type: string }
        position:
          $ref: '#/components/schemas/Position'
        currentClubId: { type: string, format: uuid, nullable: true, description: "Null if the player has exited the EPL (BR-066)." }
    Gameweek:
      type: object
      properties:
        gameweekId: { type: string, format: uuid }
        eplSeasonIdentifier: { type: string }
        number: { type: integer }
        rosterLockDeadline: { type: string, format: date-time, description: "BR-093" }
    Fixture:
      type: object
      properties:
        fixtureId: { type: string, format: uuid }
        gameweekId: { type: string, format: uuid }
        homeClubId: { type: string, format: uuid }
        awayClubId: { type: string, format: uuid }
        kickoffTime: { type: string, format: date-time }
        status:
          type: string
          enum: [Scheduled, Postponed, InProgress, Completed, Abandoned]
          description: "BR-100–BR-106"
        homeGoals: { type: integer, nullable: true }
        awayGoals: { type: integer, nullable: true }
    ClubStanding:
      type: object
      description: Real-world EPL table (BR-329–BR-331); recomputed, never Administrator-editable (BR-334).
      properties:
        clubId: { type: string, format: uuid }
        position: { type: integer }
        played: { type: integer }
        won: { type: integer }
        drawn: { type: integer }
        lost: { type: integer }
        goalsFor: { type: integer }
        goalsAgainst: { type: integer }
        goalDifference: { type: integer }
        points: { type: integer }
...